+++ src/signing/strategies/gnosis/api/gnosisEoa.ts
@@ -1,5 +1,5 @@
 import { privateKeyToAccount } from "viem/accounts";
-import { GnosisSigningStrategy } from "./gnosis";
+import { GnosisApiStrategy } from "./gnosisApi";
 import { SafeTransaction } from '@safe-global/types-kit';
 import { getEip712TxTypes } from "@safe-global/protocol-kit/dist/src/utils/eip-712/index"
 import { checkShouldSignGnosisMessage, privateKey } from "../../../../commands/prompts";
@@ -7,7 +7,7 @@ import { ICachedArg } from "../../../strategy";
 import { SavebleDocument, Transaction } from "../../../../metadata/metadataStore";
 import { TDeploy } from "../../../../metadata/schema";
 
-export class GnosisEOAStrategy extends GnosisSigningStrategy {
+export class GnosisEOAApiStrategy extends GnosisApiStrategy {
     id = "gnosis.api.eoa";
     description = "[Not Private] Gnosis SAFE - signing w/ private key using Gnosis API";
     privateKey: ICachedArg<`0x${string}`>