+++ src/signing/strategies/strategies.ts
@@ -2,13 +2,15 @@ import { SavebleDocument, Transaction } from '../../metadata/metadataStore';
 import { TDeploy } from '../../metadata/schema';
 import {Strategy} from '../strategy';
 import EOASigningStrategy from './eoa/privateKey';
-import { GnosisEOAStrategy } from './gnosis/api/gnosisEoa';
+import { GnosisOnchainStrategy } from './gnosis/onchain/onchain';
+import { GnosisEOAApiStrategy } from './gnosis/api/gnosisEoa';
 import { GnosisLedgerStrategy } from './gnosis/api/gnosisLedger';
 import { LedgerSigningStrategy } from './eoa/ledger';
 
 export const all: (new (deploy: SavebleDocument<TDeploy>, metadata: Transaction) => Strategy)[] = [
     EOASigningStrategy,
     LedgerSigningStrategy,
-    GnosisEOAStrategy,
+    GnosisEOAApiStrategy,
     GnosisLedgerStrategy,
+    GnosisOnchainStrategy
 ]